+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - removed integrated MCP server and added it as separate package @aiderdesk/mcp-server
 - added special UI elements for edited files in assistant message after smart compaction
 - added non-interactive CLI run command to execute prompts on runnning server
+- added Move to top action for task in task sidebar
 
 ## [0.70.0]